requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.36.2
CLIP @ git+https://github.com/openai/CLIP.git
faiss-cpu==1.7.4
Pillow==10.1.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
matplotlib==3.8.2
openai==1.3.5
huggingface-hub==0.19.4
grad-cam==1.4.8
pyyaml==6.0.1
opencv-python==4.8.1.78
PIL==9.5.0
tqdm==4.66.1
requests==2.31.0
scipy==1.11.4
llava @ git+https://github.com/haotian-liu/LLaVA.git
